Try to install nox with the latest available (non-prerelease) Python
`@actions/tool-cache` has `evaluateVersions` which could be used, but it
barfs on invalid semvers, so we need to pre-filter available versions
with `semver.valid` anyway.

Just use `semver.maxSatisfying` while at it. Let `@actions/tool-cache`
bring in the `semver` dep it works with. Fall back to the old behavior
in case none of the versions are valid semvers.

Refs https://github.com/daisylb/setup-nox/issues/431